Move storage changeset code into one file
